Update ob_neurobit.cpp
Allow the use of old NeurobitRuntime files, that still work with the outdated Neurobit Optima 2 and 4 devices.

As the newer NeurobitRuntime files are unable to connect to the old devices via Bluetooth for some reason.

To make use of this, you also need the "NeurobitRuntime" folder from Brainbay v2.0 or older. And replace the newer NeurobitRuntime folder with this old one.

This new ob_neurobit.cpp support both the old and new NeurobitRuntime files. So it is compatible with both the new Neurobit Optima devices and the old.