Upgrade AVERAGER
Now averager supports setting averaging window by seconds OR by number of events.

Events mean that it will average based on number of numbers it receives. So if it is set to average 10 events and it receives 10 BPM numbers, regardless of how many samples or time it took to get those, it will average those 10.
This way you can easily average the last 10 heart beats or the last 10 breathing rates etc.
Up to 1000 events!