Adds presets to Expression Evaluator
Expression Evaluator now gets a menu of common formula presets!
And a new text box for setting the name you want to the block.
And the button also deactivates when you select from the menu, as the change takes effect instantly.
It reactivates as soon as you make custom changes to either the title or formula fields yourself.